// src/app/_dto/TblHActDeactive.dto.ts
import { TbldActDeactiveCo } from "./TbldActDeactiveCo.dto";

export class TblHActDeactive {

    HActDeactive_SysID: number;
    HActDeactive_Code: string;
    HActDeactive_Name: string;
    HActDeactive_AcDe_Yn: string;
    // HActDeactive_AcDe_SysID: number;
    // HActDeactive_AcDe_Code: string;
    // HActDeactive_AcDe_Name: string;

    TbldActDeactiveCo: TbldActDeactiveCo[] = [];


}